@@ -4,9 +4,9 @@
   "description": "Averages colour on a given site/query",
   "main": "app.js",
   "scripts": {
-    "start": "node ./bin/www",
-    "test": "mocha --exit",
-    "postinstall": "bower install && tsc && coffee -c ."
+    "start": "npm run build && cross-env NODE_PATH=./ node ./bin/www",
+    "test": "cross-env NODE_ENV=test NODE_PATH=./ mocha test/**/*.js --exit",
+    "build": "./node_modules/.bin/webpack"
   },
   "repository": {
     "type": "git",
@@ -19,25 +19,38 @@
   },
   "homepage": "https://github.com/aXises/SiteAverager#readme",
   "dependencies": {
-    "@types/async": "^2.0.47",
-    "@types/jsdom": "^11.0.4",
-    "@types/mocha": "^2.2.48",
-    "@types/node": "^9.4.6",
-    "@types/request": "^2.47.0",
-    "async": "^2.6.0",
-    "body-parser": "^1.18.2",
-    "bower": "^1.8.2",
-    "coffeescript": "^2.2.1",
+    "body-parser": "^1.18.3",
     "cookie-parser": "^1.4.3",
-    "express": "^4.16.2",
+    "cross-env": "^5.2.0",
+    "express": "^4.16.4",
     "jquery": "^3.3.1",
-    "jsdom": "^11.6.2",
-    "less-middleware": "^2.2.1",
-    "mocha": "^5.0.1",
-    "morgan": "^1.9.0",
-    "pug": "^2.0.0-rc.4",
+    "jsdom": "^11.12.0",
+    "less-middleware": "^3.0.1",
+    "morgan": "^1.9.1",
+    "pug": "^2.0.3",
     "python-shell": "^0.5.0",
-    "request": "^2.83.0",
-    "typescript": "^2.7.1"
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
+    "request-promise": "^4.2.2",
+    "reset.css": "^2.0.2"
+  },
+  "devDependencies": {
+    "@types/async": "^2.0.50",
+    "@types/chai": "^4.1.7",
+    "@types/express": "^4.16.0",
+    "@types/jquery": "^3.3.28",
+    "@types/jsdom": "^11.12.0",
+    "@types/mocha": "^5.2.5",
+    "@types/node": "^9.6.41",
+    "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
+    "@types/request": "^2.48.1",
+    "@types/request-promise": "^4.1.42",
+    "chai": "^4.2.0",
+    "mocha": "^5.2.0",
+    "tslint": "^5.11.0",
+    "typescript": "^2.9.2",
+    "webpack": "^4.28.1",
+    "webpack-cli": "^3.2.1"
   }
 }
